With linear conflict, time taken to execute a problem with 50 moves has decreased by a factor of 4

+++ SlidingTileProject/src/slidingtileproject/HeuristicFunction.java
@@ -13,5 +13,5 @@ public interface HeuristicFunction {
 
     public int calculateHeuristic(State state);
 
-    public float calculateSingleHeuristic(State next);
+    public int calculateSingleHeuristic(State next);
 }
